refactors serializers and viewsets
