+++ care/facility/api/serializers/__init__.py
@@ -1,0 +1,1 @@
+TIMESTAMP_FIELDS = ('created_date', 'modified_date', 'deleted',)
