+++ care/facility/api/serializers/ambulance.py
@@ -1,40 +1,8 @@
-from django.contrib.auth import get_user_model
 from django.db import transaction
-from drf_extra_fields.geo_fields import PointField
 from rest_framework import serializers
 
-from care.facility.models import FACILITY_TYPES, AmbulanceDriver
-from care.facility.models import Facility, Ambulance
-from config.serializers import ChoiceField
-
-User = get_user_model()
-
-TIMESTAMP_FIELDS = ('created_date', 'modified_date', 'deleted',)
-
-
-class FacilitySerializer(serializers.ModelSerializer):
-    """Serializer for facility.models.Facility."""
-
-    district = ChoiceField(choices=User.DISTRICT_CHOICES)
-    facility_type = ChoiceField(choices=FACILITY_TYPES)
-    # A valid location => {
-    #     "latitude": 49.8782482189424,
-    #     "longitude": 24.452545489
-    # }
-    location = PointField(required=False)
-
-    class Meta:
-        model = Facility
-        fields = [
-            "id",
-            "name",
-            "district",
-            "facility_type",
-            "address",
-            "location",
-            "oxygen_capacity",
-            "phone_number",
-        ]
+from care.facility.api.serializers import TIMESTAMP_FIELDS
+from care.facility.models import Ambulance, AmbulanceDriver
 
 
 class AmbulanceDriverSerializer(serializers.ModelSerializer):
