+++ care/facility/api/viewsets/ambulance.py
@@ -1,44 +1,12 @@
 from django_filters import rest_framework as filters
 from rest_framework import status, serializers
 from rest_framework.decorators import action
-from rest_framework.mixins import (
-    ListModelMixin,
-    RetrieveModelMixin,
-    CreateModelMixin,
-    UpdateModelMixin,
-    DestroyModelMixin)
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.mixins import ListModelMixin
 from rest_framework.response import Response
-from rest_framework.viewsets import GenericViewSet
 
-from care.facility.api.serializers import FacilitySerializer, AmbulanceSerializer, AmbulanceDriverSerializer
-from care.facility.models import Facility, Ambulance
-
-
-class FacilityBaseViewset(CreateModelMixin, RetrieveModelMixin,
-                          UpdateModelMixin, DestroyModelMixin, GenericViewSet):
-    """Base class for all endpoints related to Faclity model."""
-
-    permission_classes = (IsAuthenticated,)
-
-
-class FacilityViewSet(FacilityBaseViewset, ListModelMixin):
-    """Viewset for facility CRUD operations."""
-
-    serializer_class = FacilitySerializer
-    queryset = Facility.objects.filter(is_active=True)
-
-    def get_queryset(self):
-        user = self.request.user
-        if user.is_superuser:
-            return self.queryset
-        return self.queryset.filter(created_by=user)
-
-    def perform_create(self, serializer):
-        serializer.save(created_by=self.request.user)
-
-    def perform_update(self, serializer):
-        serializer.save(created_by=self.request.user)
+from care.facility.api.serializers.ambulance import AmbulanceSerializer, AmbulanceDriverSerializer
+from care.facility.api.viewsets import FacilityBaseViewset
+from care.facility.models import Ambulance
 
 
 class AmbulanceFilterSet(filters.FilterSet):
